Refactor salescommand package to sales
