+++ src/main/java/seedu/address/logic/commands/sales/RecordSalesCommand.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands.salescommands;
+package seedu.address.logic.commands.sales;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_DATE;
