+++ src/test/java/seedu/address/logic/parser/sales/RecordSalesCommandParserTest.java
@@ -22,7 +22,7 @@ import static seedu.address.testutil.salesrecords.TypicalRecords.RECORD_ONE;
 
 import org.junit.Test;
 
-import seedu.address.logic.commands.salescommands.RecordSalesCommand;
+import seedu.address.logic.commands.sales.RecordSalesCommand;
 import seedu.address.logic.parser.salescommandsparser.RecordSalesCommandParser;
 import seedu.address.model.salesrecord.Date;
 import seedu.address.model.salesrecord.ItemName;
