+++ src/test/java/seedu/address/testutil/salesrecords/RecordUtil.java
@@ -4,7 +4,7 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_ITEM_NAME;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_ITEM_PRICE;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_QUANTITY_SOLD;
 
-import seedu.address.logic.commands.salescommands.RecordSalesCommand;
+import seedu.address.logic.commands.sales.RecordSalesCommand;
 import seedu.address.model.salesrecord.SalesRecord;
 /**
  * A utility class for {@code SalesRecord}.
